+++ matrix-sdk-android-rx/src/main/java/im/vector/matrix/rx/RxRoom.kt
@@ -57,8 +57,9 @@ class RxRoom(private val room: Room) {
         room.loadRoomMembersIfNeeded(MatrixCallbackSingle(it)).toSingle(it)
     }
 
-    fun joinRoom(viaServers: List<String> = emptyList()): Single<Unit> = Single.create {
-        room.join(viaServers, MatrixCallbackSingle(it)).toSingle(it)
+    fun joinRoom(reason: String? = null,
+                 viaServers: List<String> = emptyList()): Single<Unit> = Single.create {
+        room.join(reason, viaServers, MatrixCallbackSingle(it)).toSingle(it)
     }
 
     fun liveEventReadReceipts(eventId: String): Observable<List<ReadReceipt>> {