+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/room/RoomService.kt
@@ -30,14 +30,17 @@ interface RoomService {
     /**
      * Create a room asynchronously
      */
-    fun createRoom(createRoomParams: CreateRoomParams, callback: MatrixCallback<String>): Cancelable
+    fun createRoom(createRoomParams: CreateRoomParams,
+                   callback: MatrixCallback<String>): Cancelable
 
     /**
      * Join a room by id
      * @param roomId the roomId of the room to join
+     * @param reason optional reason for joining the room
      * @param viaServers the servers to attempt to join the room through. One of the servers must be participating in the room.
      */
     fun joinRoom(roomId: String,
+                 reason: String? = null,
                  viaServers: List<String> = emptyList(),
                  callback: MatrixCallback<Unit>): Cancelable
 
@@ -69,5 +72,6 @@ interface RoomService {
     /**
      * Mark all rooms as read
      */
-    fun markAllAsRead(roomIds: List<String>, callback: MatrixCallback<Unit>): Cancelable
+    fun markAllAsRead(roomIds: List<String>,
+                      callback: MatrixCallback<Unit>): Cancelable
 }